Implement partial apply, but not work for generic.
For generics, partial application is difficult in go lang.
So I decide just generate code as is even though it is error in
non-resolve type parameter cases in go lang.
We can use it in pipe line scenario because in that case type is
resolved before generating go code.
But it is difficult to test it in current stage, so I just keep my code
and then test it after implementing pipe operator.

[go - Generic function, can't be defined in form of anonymous? - Stack
Overflow](https://stackoverflow.com/questions/75915359/generic-function-cant-be-defined-in-form-of-anonymous)